+++ pick-front/src/components/post/SinglePost.vue
@@ -44,7 +44,7 @@
 
 <script setup>
     import SearchBox from '@/components/common/SearchBox.vue';
-    import Pagination from '@/components/Pagination.vue'
+    import Pagination from '@/components/common/Pagination.vue';
     import List from '@/components/post/List.vue';
     import postList from '@/json/post_list.json';
